consumer 模块监听消息处理优化， 1.会员模块采用了策略模式 + 门面路由模式 优化会员模块（类爆炸） 2.商品模块使用了类似springMVC dispatcher 的方法测试模式，彻底消灭if else case 以上2种优化模式作为对比，实际生产中倾向于第二种
