+++ consumer/src/main/java/cn/lili/annotation/GoodsMQTagRoute.java
@@ -1,0 +1,13 @@
+package cn.lili.annotation;
+
+import cn.lili.rocketmq.tags.GoodsTagsEnum;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface GoodsMQTagRoute {
+
+    GoodsTagsEnum value();
+}
